code clean up + added comments

+++ src/Renderer/swapchain_resources.cpp
@@ -11,12 +11,22 @@ void Renderer::initSwapChainResources()
 {
   qDebug("initSwapChainResources");
 
-  // Projection matrix
-  m_proj = m_vkWindow->clipCorrectionMatrix(); // adjust for Vulkan-OpenGL clip space differences
+  /**
+   * NOTE:
+   *
+   * The vertex data and the projection matrix assume OpenGL. With
+   * Vulkan Y is negated in clip space and the near/far plane is at 0/1 instead
+   * of -1/1. These will be corrected for by an extra transformation when
+   * calculating the model-view-projection matrix.
+   *
+   * Y: up, front: CCW
+   *
+   */
+  m_proj = m_vkWindow->clipCorrectionMatrix();
   auto windowSize = m_vkWindow->swapChainImageSize();
   m_proj.perspective(
     45.0f,
-    windowSize.width() / (float) windowSize.height(),
+    (float) windowSize.width() / (float) windowSize.height(),
     0.01f,
     1000.0f
   );
